Packing Weight Length Double

+++ Com.Danliris.Service.Production.Lib/ViewModels/Packing/PackingDetailViewModel.cs
@@ -13,8 +13,8 @@ namespace Com.Danliris.Service.Finishing.Printing.Lib.ViewModels.Packing
         public string Lot { get; set; }
         [MaxLength(100)]
         public string Grade { get; set; }
-        public int Weight { get; set; }
-        public int Length { get; set; }
+        public double Weight { get; set; }
+        public double Length { get; set; }
         public int Quantity { get; set; }
         [MaxLength(500)]
         public string Remark { get; set; }
